Rename to oapi_validate_request to distinguish for file for response Move shared code into types and utils files Add proper types for context keys to get rid of linter warning
